+++ data/types/basal/basal_test.go
@@ -5,7 +5,6 @@ import (
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
 
-	"github.com/tidepool-org/platform/app"
 	"github.com/tidepool-org/platform/data/context"
 	"github.com/tidepool-org/platform/data/factory"
 	"github.com/tidepool-org/platform/data/normalizer"
@@ -13,6 +12,7 @@ import (
 	testData "github.com/tidepool-org/platform/data/test"
 	"github.com/tidepool-org/platform/data/types/basal"
 	"github.com/tidepool-org/platform/data/validator"
+	"github.com/tidepool-org/platform/id"
 	"github.com/tidepool-org/platform/log"
 	"github.com/tidepool-org/platform/pointer"
 	"github.com/tidepool-org/platform/service"
@@ -28,7 +28,7 @@ func NewMeta(deliveryType string) interface{} {
 func NewTestBasal(sourceTime interface{}, sourceDeliveryType interface{}) *basal.Basal {
 	testBasal := &basal.Basal{}
 	testBasal.Init()
-	testBasal.DeviceID = pointer.String(app.NewID())
+	testBasal.DeviceID = pointer.String(id.New())
 	if value, ok := sourceTime.(string); ok {
 		testBasal.Time = pointer.String(value)
 	}
@@ -181,8 +181,8 @@ var _ = Describe("Basal", func() {
 				var deviceID string
 
 				BeforeEach(func() {
-					userID = app.NewID()
-					deviceID = app.NewID()
+					userID = id.New()
+					deviceID = id.New()
 					testBasal.UserID = userID
 					testBasal.DeviceID = &deviceID
 					testBasal.Time = pointer.String("2016-09-06T13:45:58-07:00")